Split module system loading helpers
