+++ src/module_system/mod.rs
@@ -1,11 +1,16 @@
 use std::collections::{HashMap, HashSet};
 use std::path::{Path, PathBuf};
 
-use crate::ast::{Declaration, Program};
+use crate::ast::Program;
 use crate::error::{CompileError, FileTable, Span};
 use crate::lexer;
 use crate::parser;
-use crate::resolver::{Resolver, SymbolTable};
+use crate::resolver::SymbolTable;
+
+mod graph_loading;
+mod import_resolution;
+
+use import_resolution::find_stdlib_root;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct ModuleId(pub u32);
@@ -170,496 +175,6 @@ impl ModuleSystem {
         Ok(program)
     }
 
-    /// Load a module graph with validated import bindings.
-    ///
-    /// Unlike `load_with_imports`, this does not merge imported declarations
-    /// into the entry module AST. It records imports as bindings between module
-    /// IDs so resolver/typechecker integration can move away from AST cloning.
-    pub fn load_module_graph(
-        &mut self,
-        path: &Path,
-        files: &mut FileTable,
-    ) -> Result<ResolvedModuleGraph, Vec<CompileError>> {
-        let mut graph = ResolvedModuleGraph {
-            entry: ModuleId(0),
-            modules: HashMap::new(),
-            paths: HashMap::new(),
-        };
-        let mut loading = HashSet::new();
-        let entry = self.load_graph_module(path, files, &mut graph, &mut loading)?;
-        graph.entry = entry;
-        Ok(graph)
-    }
-
-    /// Walk Import declarations in `program` and load their dependencies.
-    ///
-    /// Import routing:
-    /// - `{ io } = std` or `{ io } = std.io` — stdlib, skip (handled by codegen)
-    /// - `{ cast } = @builtin` — intrinsic, skip
-    /// - `{ add } = math` — resolve `math.zen` relative to `base_dir`
-    /// - `{ Foo } = utils.math` — resolve `utils/math.zen` relative to `base_dir`
-    fn resolve_imports(
-        &mut self,
-        program: &mut Program,
-        base_dir: &Path,
-        files: &mut FileTable,
-    ) -> Result<(), Vec<CompileError>> {
-        // Collect imports first to avoid borrow conflict
-        let imports: Vec<(Vec<String>, Vec<String>, Span)> = program
-            .declarations
-            .iter()
-            .filter_map(|decl| match decl {
-                Declaration::Import {
-                    names,
-                    module_path,
-                    span,
-                } => Some((names.clone(), module_path.clone(), *span)),
-                _ => None,
-            })
-            .collect();
-
-        let mut imported_decls: Vec<Declaration> = Vec::new();
-
-        for (names, module_path, span) in imports {
-            if module_path.is_empty() {
-                return Err(vec![CompileError::Resolution(
-                    "empty import path".into(),
-                    Some(span),
-                )]);
-            }
-
-            let first = &module_path[0];
-
-            if first == "std" || first == "@std" {
-                if module_path.len() == 1 {
-                    // Root prelude imports are still handled by codegen/runtime.
-                    continue;
-                }
-
-                let Some(file_path) = self.resolve_stdlib_file_path(&module_path[1..])? else {
-                    return Err(vec![CompileError::Resolution(
-                        format!("cannot find stdlib module '{}'", module_path.join(".")),
-                        Some(span),
-                    )]);
-                };
-
-                let dep_program = self.load_with_imports(&file_path, files)?;
-                self.collect_imported_declarations(
-                    &dep_program,
-                    &names,
-                    &module_path.join("."),
-                    span,
-                    &mut imported_decls,
-                )?;
-                continue;
-            }
-
-            if first == "@builtin" {
-                // Builtin intrinsics — nothing to load
-                continue;
-            }
-
-            let file_path = self.local_import_file_path(base_dir, &module_path, span)?;
-            self.reject_duplicate_requested_imports(&names, &module_path, span)?;
-
-            let dep_program = self.load_with_imports(&file_path, files)?;
-            self.collect_imported_declarations(
-                &dep_program,
-                &names,
-                &module_path.join("."),
-                span,
-                &mut imported_decls,
-            )?;
-        }
-
-        // Append imported declarations to the program
-        program.declarations.extend(imported_decls);
-        Ok(())
-    }
-
-    fn local_import_file_path(
-        &self,
-        base_dir: &Path,
-        module_path: &[String],
-        span: Span,
-    ) -> Result<PathBuf, Vec<CompileError>> {
-        // Local file import: resolve module_path segments as a file path
-        // e.g. ["math"] -> math.zen, ["utils", "math"] -> utils/math.zen
-        let rel_path: PathBuf = module_path.iter().collect();
-        let mut file_path = base_dir.join(&rel_path);
-        if file_path.extension().is_none() {
-            file_path.set_extension("zen");
-        }
-
-        if !file_path.exists() {
-            return Err(vec![CompileError::Resolution(
-                format!(
-                    "cannot find imported module '{}' (looked for {})",
-                    module_path.join("."),
-                    file_path.display()
-                ),
-                Some(span),
-            )]);
-        }
-
-        Ok(file_path)
-    }
-
-    fn reject_duplicate_requested_imports(
-        &self,
-        names: &[String],
-        module_path: &[String],
-        span: Span,
-    ) -> Result<(), Vec<CompileError>> {
-        let mut requested_names = HashSet::new();
-        for name in names {
-            if !requested_names.insert(name.as_str()) {
-                return Err(vec![CompileError::Resolution(
-                    format!(
-                        "duplicate import '{}' from module '{}'",
-                        name,
-                        module_path.join(".")
-                    ),
-                    Some(span),
-                )]);
-            }
-        }
-        Ok(())
-    }
-
-    fn collect_imported_declarations(
-        &self,
-        dep_program: &Program,
-        names: &[String],
-        module_name: &str,
-        import_span: Span,
-        imported_decls: &mut Vec<Declaration>,
-    ) -> Result<(), Vec<CompileError>> {
-        for name in names {
-            let mut found_private = false;
-            let mut found_public = false;
-
-            for decl in &dep_program.declarations {
-                if decl.name() == Some(name.as_str()) {
-                    if decl.is_public() {
-                        found_public = true;
-                        imported_decls.push(decl.clone());
-                    } else {
-                        found_private = true;
-                    }
-                }
-
-                // Importing a public type also brings its public methods into scope.
-                if let Declaration::Method {
-                    type_name, public, ..
-                } = decl
-                {
-                    if type_name == name && *public {
-                        imported_decls.push(decl.clone());
-                    }
-                }
-            }
-
-            if !found_public {
-                if found_private {
-                    return Err(vec![CompileError::Resolution(
-                        format!(
-                            "symbol '{}' in module '{}' is not exported",
-                            name, module_name
-                        ),
-                        Some(import_span),
-                    )]);
-                }
-                return Err(vec![CompileError::Resolution(
-                    format!("module '{}' does not export '{}'", module_name, name),
-                    Some(import_span),
-                )]);
-            }
-        }
-
-        Ok(())
-    }
-
-    fn load_graph_module(
-        &mut self,
-        path: &Path,
-        files: &mut FileTable,
-        graph: &mut ResolvedModuleGraph,
-        loading: &mut HashSet<PathBuf>,
-    ) -> Result<ModuleId, Vec<CompileError>> {
-        let canonical = path.canonicalize().map_err(|e| {
-            vec![CompileError::Internal(format!(
-                "cannot resolve path {}: {}",
-                path.display(),
-                e
-            ))]
-        })?;
-        let key = canonical.display().to_string();
-
-        if loading.contains(&canonical) {
-            return Err(vec![CompileError::Resolution(
-                format!("circular import detected: {}", canonical.display()),
-                None,
-            )]);
-        }
-
-        if let Some(id) = graph.paths.get(&key) {
-            return Ok(*id);
-        }
-
-        loading.insert(canonical.clone());
-
-        let program = self.load_file(path, files)?;
-        let symbols = Resolver::new()
-            .resolve_program(&program)
-            .map_err(|diagnostics| {
-                diagnostics
-                    .into_iter()
-                    .map(|diagnostic| CompileError::Resolution(diagnostic.message, diagnostic.span))
-                    .collect::<Vec<_>>()
-            })?;
-        let info = self.register_module_info(&key, &canonical);
-        let id = info.id;
-        graph.paths.insert(key.clone(), id);
-
-        let base_dir = canonical
-            .parent()
-            .unwrap_or_else(|| Path::new("."))
-            .to_path_buf();
-        let imports = self.resolve_graph_imports(&program, &base_dir, files, graph, loading)?;
-
-        loading.remove(&canonical);
-        graph.modules.insert(
-            id,
-            ResolvedModule {
-                info,
-                program,
-                imports,
-                symbols,
-            },
-        );
-        Ok(id)
-    }
-
-    fn resolve_graph_imports(
-        &mut self,
-        program: &Program,
-        base_dir: &Path,
-        files: &mut FileTable,
-        graph: &mut ResolvedModuleGraph,
-        loading: &mut HashSet<PathBuf>,
-    ) -> Result<Vec<ImportBinding>, Vec<CompileError>> {
-        let imports: Vec<(Vec<String>, Vec<String>, Span)> = program
-            .declarations
-            .iter()
-            .filter_map(|decl| match decl {
-                Declaration::Import {
-                    names,
-                    module_path,
-                    span,
-                } => Some((names.clone(), module_path.clone(), *span)),
-                _ => None,
-            })
-            .collect();
-
-        let mut bindings = Vec::new();
-
-        for (names, module_path, span) in imports {
-            if module_path.is_empty() {
-                return Err(vec![CompileError::Resolution(
-                    "empty import path".into(),
-                    Some(span),
-                )]);
-            }
-
-            self.reject_duplicate_requested_imports(&names, &module_path, span)?;
-
-            let first = &module_path[0];
-            if first == "@builtin"
-                || ((first == "std" || first == "@std") && module_path.len() == 1)
-            {
-                continue;
-            }
-
-            let file_path = if first == "std" || first == "@std" {
-                self.resolve_stdlib_file_path(&module_path[1..])?
-                    .ok_or_else(|| {
-                        vec![CompileError::Resolution(
-                            format!("cannot find stdlib module '{}'", module_path.join(".")),
-                            Some(span),
-                        )]
-                    })?
-            } else {
-                self.local_import_file_path(base_dir, &module_path, span)?
-            };
-
-            let dep_id = self.load_graph_module(&file_path, files, graph, loading)?;
-            let dep_module = graph
-                .module(dep_id)
-                .expect("graph module exists immediately after load");
-            self.collect_import_bindings(
-                dep_module,
-                &names,
-                &module_path.join("."),
-                span,
-                &mut bindings,
-            )?;
-        }
-
-        Ok(bindings)
-    }
-
-    fn collect_import_bindings(
-        &self,
-        dep_module: &ResolvedModule,
-        names: &[String],
-        module_name: &str,
-        import_span: Span,
-        bindings: &mut Vec<ImportBinding>,
-    ) -> Result<(), Vec<CompileError>> {
-        for name in names {
-            let mut found_private = false;
-            let mut found_public = false;
-
-            for decl in &dep_module.program.declarations {
-                if decl.name() == Some(name.as_str()) {
-                    if decl.is_public() {
-                        found_public = true;
-                    } else {
-                        found_private = true;
-                    }
-                }
-            }
-
-            if found_public {
-                bindings.push(ImportBinding {
-                    local_name: name.clone(),
-                    source_module: dep_module.info.id,
-                    source_symbol: name.clone(),
-                    span: import_span,
-                });
-                continue;
-            }
-
-            if found_private {
-                return Err(vec![CompileError::Resolution(
-                    format!(
-                        "symbol '{}' in module '{}' is not exported",
-                        name, module_name
-                    ),
-                    Some(import_span),
-                )]);
-            }
-
-            return Err(vec![CompileError::Resolution(
-                format!("module '{}' does not export '{}'", module_name, name),
-                Some(import_span),
-            )]);
-        }
-
-        Ok(())
-    }
-
-    /// Resolve an import module path to a filesystem path and load it.
-    /// e.g. ["std"] or ["std", "io"] or ["@std", "io"]
-    pub fn resolve_import(
-        &mut self,
-        module_path: &[String],
-        files: &mut FileTable,
-    ) -> Result<Program, Vec<CompileError>> {
-        if module_path.is_empty() {
-            return Err(vec![CompileError::Resolution(
-                "empty module path".into(),
-                None,
-            )]);
-        }
-
-        let first = &module_path[0];
-
-        // @std or std prefix → look in stdlib
-        if first == "@std" || first == "std" {
-            return self.resolve_stdlib_import(&module_path[1..], files);
-        }
-
-        // @builtin → return empty program (intrinsics)
-        if first == "@builtin" {
-            return Ok(Program {
-                declarations: Vec::new(),
-                file_id: 0,
-            });
-        }
-
-        Err(vec![CompileError::Resolution(
-            format!("unknown module: {}", module_path.join(".")),
-            None,
-        )])
-    }
-
-    fn resolve_stdlib_import(
-        &mut self,
-        sub_path: &[String],
-        files: &mut FileTable,
-    ) -> Result<Program, Vec<CompileError>> {
-        if sub_path.is_empty() {
-            // `{ io } = std` — load the stdlib prelude or nothing
-            return Ok(Program {
-                declarations: Vec::new(),
-                file_id: 0,
-            });
-        }
-
-        if let Some(file_path) = self.resolve_stdlib_file_path(sub_path)? {
-            return self.load_file(&file_path, files);
-        }
-
-        // Return empty for now — many stdlib modules are stubs
-        Ok(Program {
-            declarations: Vec::new(),
-            file_id: 0,
-        })
-    }
-
-    fn resolve_stdlib_file_path(
-        &self,
-        sub_path: &[String],
-    ) -> Result<Option<PathBuf>, Vec<CompileError>> {
-        let root = match &self.stdlib_root {
-            Some(r) => r.clone(),
-            None => {
-                return Err(vec![CompileError::Resolution(
-                    "stdlib not found".into(),
-                    None,
-                )]);
-            }
-        };
-
-        if sub_path.is_empty() {
-            return Ok(None);
-        }
-
-        let mut dir = root;
-        for seg in sub_path {
-            dir.push(seg);
-        }
-
-        let file_path = dir.with_extension("zen");
-        if file_path.exists() {
-            return Ok(Some(file_path));
-        }
-
-        let mod_path = dir.join("mod.zen");
-        if mod_path.exists() {
-            return Ok(Some(mod_path));
-        }
-
-        let parent_file = dir.with_extension("zen");
-        if parent_file.exists() {
-            return Ok(Some(parent_file));
-        }
-
-        Ok(None)
-    }
-
     pub fn modules(&self) -> &HashMap<String, Program> {
         &self.modules
     }
@@ -701,38 +216,6 @@ impl ModuleSystem {
             PackageId(0)
         }
     }
-}
-
-/// Find the stdlib root by looking for a `stdlib/` directory.
-fn find_stdlib_root() -> Option<PathBuf> {
-    // Try relative to current exe
-    if let Ok(exe) = std::env::current_exe() {
-        if let Some(dir) = exe.parent() {
-            // Check alongside the binary
-            let stdlib = dir.join("stdlib");
-            if stdlib.is_dir() {
-                return Some(stdlib);
-            }
-            // Check one level up (target/debug -> project root)
-            if let Some(parent) = dir.parent() {
-                if let Some(grandparent) = parent.parent() {
-                    let stdlib = grandparent.join("stdlib");
-                    if stdlib.is_dir() {
-                        return Some(stdlib);
-                    }
-                }
-            }
-        }
-    }
-
-    // Try current working directory
-    let cwd = std::env::current_dir().ok()?;
-    let stdlib = cwd.join("stdlib");
-    if stdlib.is_dir() {
-        return Some(stdlib);
-    }
-
-    None
 }
 
 // ── Tests ─────────────────────────────────────────────────────────
